WIP: fixup! drivers: input: touchscreen: Import FT touchcscreen driver from DS
This is some preparation for rewriting the driver:
- Drop downstream pinctrl, point report check, gesture, pen, ESD, drm
notifier implementation
- Drop unused code
- Drop old firmware files
- Switch to tabs in main files
- Refactor regulators & reset gpio implementation
- Drop FTS macros & use kernel log func calls
- And other minor changes...

Co-developed-by: Eugene Lepshy <fekz115@gmail.com>
