+++ drivers/input/touchscreen/Makefile
@@ -118,5 +118,5 @@ obj-$(CONFIG_TOUCHSCREEN_IQS5XX)	+= iqs5xx.o
 obj-$(CONFIG_TOUCHSCREEN_IQS7211)	+= iqs7211.o
 obj-$(CONFIG_TOUCHSCREEN_ZINITIX)	+= zinitix.o
 obj-$(CONFIG_TOUCHSCREEN_HIMAX_HX83112B)	+= himax_hx83112b.o
-obj-$(CONFIG_TOUCHSCREEN_FTS)			+= focaltech_touch/
+obj-$(CONFIG_TOUCHSCREEN_FTS_CORE)	+= focaltech_touch/
 obj-$(CONFIG_TOUCHSCREEN_FOCALTECH_FT3658U)	+= focaltech_touch_ft3658u/
