+++ drivers/input/touchscreen/focaltech_touch/Kconfig
@@ -1,15 +1,18 @@
-#
 # Focaltech Touchscreen driver configuration
-#
 
-config TOUCHSCREEN_FTS
-    tristate "Focaltech Touchscreen"
-    default n
-    help
-      Say Y here if you have Focaltech touch panel.
-      If unsure, say N.
+config TOUCHSCREEN_FTS_CORE
+	tristate
 
-config TOUCHSCREEN_FTS_DIRECTORY
-    string "Focaltech ts directory name"
-    default "focaltech_touch"
-    depends on TOUCHSCREEN_FTS
+config TOUCHSCREEN_FTS_SPI
+	tristate "Focaltech SPI touchscreen"
+	depends on SPI_MASTER
+	select REGMAP
+	select TOUCHSCREEN_FTS_CORE
+	help
+	  Say Y here if you have a Focaltech IC connected to
+	  your system via SPI.
+
+	  If unsure, say N.
+
+	  To compile this driver as a module, choose M here: the
+	  module will be called focaltech_spi.
