--- drivers/input/touchscreen/focaltech_touch/focaltech_common.h
@@ -1,169 +1,0 @@
-/*
- * FocalTech fts TouchScreen driver.
- *
- * Copyright (c) 2012-2020, Focaltech Ltd. All rights reserved.
- *
- * This software is licensed under the terms of the GNU General Public
- * License version 2, as published by the Free Software Foundation, and
- * may be copied, distributed, and modified under those terms.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- */
-
-#ifndef __LINUX_FOCALTECH_COMMON_H__
-#define __LINUX_FOCALTECH_COMMON_H__
-
-#define FTS_CHIP_TYPE 0x3680008A
-#define FTS_AUTO_UPGRADE_EN 1
-#define FTS_GET_MODULE_NUM 0
-
-/*
- * module_id: mean vendor_id generally, also maybe gpio or lcm_id...
- * If means vendor_id, the FTS_MODULE_ID = PANEL_ID << 8 + VENDOR_ID
- * FTS_GET_MODULE_NUM == 0/1, no check module id, you may ignore them
- * FTS_GET_MODULE_NUM >= 2, compatible with FTS_MODULE2_ID
- * FTS_GET_MODULE_NUM >= 3, compatible with FTS_MODULE3_ID
- */
-#define FTS_MODULE_ID 0x0000
-#define FTS_MODULE2_ID 0x0000
-#define FTS_MODULE3_ID 0x0000
-
-/*
- * Need set the following when get firmware via firmware_request()
- * For example: if module'vendor is tianma,
- * #define FTS_MODULE_NAME "tianma"
- * then file_name will be "focaltech_ts_fw_tianma"
- * You should rename fw to "focaltech_ts_fw_tianma", and push it into
- * etc/firmware or by customers
- */
-#define FTS_MODULE_NAME		""
-#define FTS_MODULE2_NAME	""
-#define FTS_MODULE3_NAME	""
-
-/*
- * FW.i file for auto upgrade, you must replace it with your own
- * define your own fw_file, the sample one to be replaced is invalid
- * NOTE: if FTS_GET_MODULE_NUM > 1, it's the fw corresponding with FTS_VENDOR_ID
- */
-#define FTS_UPGRADE_FW_FILE \
-	"include/firmware/FT3680_WXN_M146_V27_D01_20220706_app.i"
-
-/*
- * if FTS_GET_MODULE_NUM >= 2, fw corrsponding with FTS_VENDOR_ID2
- * define your own fw_file, the sample one is invalid
- */
-#define FTS_UPGRADE_FW2_FILE	"include/firmware/fw_sample.i"
-
-/*
- * if FTS_GET_MODULE_NUM >= 3, fw corrsponding with FTS_VENDOR_ID3
- * define your own fw_file, the sample one is invalid
- */
-#define FTS_UPGRADE_FW3_FILE	"include/firmware/fw_sample.i"
-
-/*********************************************************/
-
-/*****************************************************************************
-* Macro definitions using #define
-*****************************************************************************/
-#define FTS_DRIVER_VERSION	"Focaltech V3.4 20211214"
-
-#define BYTE_OFF_0(x)		(u8)((x) & 0xFF)
-#define BYTE_OFF_8(x)		(u8)(((x) >> 8) & 0xFF)
-#define BYTE_OFF_16(x)		(u8)(((x) >> 16) & 0xFF)
-#define BYTE_OFF_24(x)		(u8)(((x) >> 24) & 0xFF)
-#define FLAGBIT(x)		(0x00000001 << (x))
-#define FLAGBITS(x, y)		((0xFFFFFFFF >> (32 - (y) - 1)) & (0xFFFFFFFF << (x)))
-
-#define FLAG_ICSERIALS_LEN	8
-#define FLAG_HID_BIT		10
-#define FLAG_IDC_BIT		11
-
-#define IC_SERIALS		(FTS_CHIP_TYPE & FLAGBITS(0, FLAG_ICSERIALS_LEN-1))
-#define IC_TO_SERIALS(x)	((x) & FLAGBITS(0, FLAG_ICSERIALS_LEN-1))
-#define FTS_CHIP_IDC		((FTS_CHIP_TYPE & FLAGBIT(FLAG_IDC_BIT)) == FLAGBIT(FLAG_IDC_BIT))
-#define FTS_HID_SUPPORTTED	((FTS_CHIP_TYPE & FLAGBIT(FLAG_HID_BIT)) == FLAGBIT(FLAG_HID_BIT))
-
-#define FTS_MAX_CHIP_IDS	8
-
-#define FTS_CHIP_TYPE_MAPPING \
-				{ { 0x8A, 0x56, 0x62, 0x56, 0x62, 0x56, 0xE2, 0x00, 0x00 } }
-
-#define FILE_NAME_LENGTH			128
-#define ENABLE					1
-#define DISABLE					0
-#define VALID					1
-#define INVALID					0
-#define FTS_CMD_START1				0x55
-#define FTS_CMD_START2				0xAA
-#define FTS_CMD_START_DELAY			12
-#define FTS_CMD_READ_ID				0x90
-#define FTS_CMD_READ_ID_LEN			4
-#define FTS_CMD_READ_ID_LEN_INCELL		1
-#define FTS_CMD_READ_INFO			0x5E
-
-/*register address*/
-#define FTS_REG_INT_CNT				0x8F
-#define FTS_REG_FLOW_WORK_CNT			0x91
-#define FTS_REG_WORKMODE			0x00
-#define FTS_REG_WORKMODE_FACTORY_VALUE		0x40
-#define FTS_REG_WORKMODE_WORK_VALUE		0x00
-#define FTS_REG_CHIP_ID				0xA3
-#define FTS_REG_CHIP_ID2			0x9F
-#define FTS_REG_POWER_MODE			0xA5
-#define FTS_REG_POWER_MODE_SLEEP		0x03
-#define FTS_REG_FW_VER				0xA6
-#define FTS_REG_VENDOR_ID			0xA8
-#define FTS_REG_LCD_BUSY_NUM			0xAB
-#define FTS_REG_FACE_DEC_MODE_EN		0xB0
-#define FTS_REG_FACTORY_MODE_DETACH_FLAG	0xB4
-#define FTS_REG_FACE_DEC_MODE_STATUS		0x01
-#define FTS_REG_IDE_PARA_VER_ID			0xB5
-#define FTS_REG_IDE_PARA_STATUS			0xB6
-#define FTS_REG_GLOVE_MODE_EN			0xC0
-#define FTS_REG_COVER_MODE_EN			0xC1
-#define FTS_REG_CHARGER_MODE_EN			0x8B
-#define FTS_REG_MODULE_ID			0xE3
-#define FTS_REG_LIC_VER				0xE4
-#define FACTORY_REG_OPEN_ADDR			0xCF
-#define FACTORY_REG_OPEN_ADDR_FOD		0x02
-#define FTS_REG_EDGE_MODE_EN			0x8C
-
-#define FTS_SYSFS_ECHO_ON(buf)		(buf[0] == '1' || buf[0] == '2')
-#define FTS_SYSFS_ECHO_OFF(buf)		(buf[0] == '0')
-
-#define kfree_safe(pbuf)             \
-	do {                         \
-		if (pbuf) {          \
-			kfree(pbuf); \
-			pbuf = NULL; \
-		}                    \
-	} while (0)
-
-struct ft_chip_t {
-	u16 type;
-	u8 chip_idh;
-	u8 chip_idl;
-	u8 rom_idh;
-	u8 rom_idl;
-	u8 pb_idh;
-	u8 pb_idl;
-	u8 bl_idh;
-	u8 bl_idl;
-};
-
-struct ft_chip_id_t {
-	u16 type;
-	u16 chip_ids[FTS_MAX_CHIP_IDS];
-};
-
-struct ts_ic_info {
-	bool is_incell;
-	bool hid_supported;
-	struct ft_chip_t ids;
-	struct ft_chip_id_t cid;
-};
-
-#endif /* __LINUX_FOCALTECH_COMMON_H__ */
