+++ drivers/input/touchscreen/focaltech_touch/focaltech_core.c
@@ -19,11 +19,8 @@
 #include <linux/of_device.h>
 #include <linux/of_gpio.h>
 #include <linux/of_irq.h>
-#include <linux/notifier.h>
-#include <linux/fb.h>
-#include <drm/drm_panel.h>
 
-#include "focaltech_core.h"
+#include "focaltech.h"
 
 #define FTS_DRIVER_NAME		"fts_ts"
 #define INTERVAL_READ_REG	200 /* unit:ms */
@@ -84,23 +81,6 @@ int fts_wait_tp_to_valid(void)
 	} while ((cnt * INTERVAL_READ_REG) < TIMEOUT_READ_REG);
 
 	return -EIO;
-}
-
-/*****************************************************************************
-*  Name: fts_tp_state_recovery
-*  Brief: Need execute this function when reset
-*  Input:
-*  Output:
-*  Return:
-*****************************************************************************/
-void fts_tp_state_recovery(struct fts_ts_data *ts_data)
-{
-	/* wait tp stable */
-	fts_wait_tp_to_valid();
-	/* recover TP charger state 0x8B */
-	/* recover TP glove state 0xC0 */
-	/* recover TP cover state 0xC1 */
-	fts_ex_mode_recovery(ts_data);
 }
 
 int fts_request_handle_reset(struct fts_ts_data *ts_data, int hdelayms)
@@ -172,48 +152,6 @@ void fts_hid2std(void)
 	}
 }
 
-static int fts_match_cid(struct fts_ts_data *ts_data, u16 type, u8 id_h,
-			 u8 id_l, bool force)
-{
-#ifdef FTS_CHIP_ID_MAPPING
-	u32 i = 0;
-	u32 j = 0;
-	struct ft_chip_id_t chip_id_list[] = FTS_CHIP_ID_MAPPING;
-	u32 cid_entries = sizeof(chip_id_list) / sizeof(struct ft_chip_id_t);
-	u16 id = (id_h << 8) + id_l;
-
-	memset(&ts_data->ic_info.cid, 0, sizeof(struct ft_chip_id_t));
-	for (i = 0; i < cid_entries; i++) {
-		if (!force && (type == chip_id_list[i].type))
-			break;
-		else if (force && (type == chip_id_list[i].type)) {
-			dev_info(dev, "match cid,type:0x%x",
-				 (int)chip_id_list[i].type);
-			ts_data->ic_info.cid = chip_id_list[i];
-			return 0;
-		}
-	}
-
-	if (i >= cid_entries)
-		return -ENODATA;
-
-	for (j = 0; j < FTS_MAX_CHIP_IDS; j++) {
-		if (id == chip_id_list[i].chip_ids[j]) {
-			dev_dbg(dev, "cid:%x==%x", id,
-				chip_id_list[i].chip_ids[j]);
-			dev_info(dev, "match cid,type:0x%x",
-				 (int)chip_id_list[i].type);
-			ts_data->ic_info.cid = chip_id_list[i];
-			return 0;
-		}
-	}
-
-	return -ENODATA;
-#else
-	return -EINVAL;
-#endif
-}
-
 static int fts_get_chip_types(struct fts_ts_data *ts_data, u8 id_h, u8 id_l,
 			      bool fw_valid)
 {
@@ -230,9 +168,7 @@ static int fts_get_chip_types(struct fts_ts_data *ts_data, u8 id_h, u8 id_l,
 	for (i = 0; i < ctype_entries; i++) {
 		if (VALID == fw_valid) {
 			if (((id_h == ctype[i].chip_idh) &&
-			     (id_l == ctype[i].chip_idl)) ||
-			    (!fts_match_cid(ts_data, ctype[i].type, id_h, id_l,
-					    0)))
+			     (id_l == ctype[i].chip_idl)))
 				break;
 		} else {
 			if (((id_h == ctype[i].rom_idh) &&
@@ -249,7 +185,6 @@ static int fts_get_chip_types(struct fts_ts_data *ts_data, u8 id_h, u8 id_l,
 	if (i >= ctype_entries)
 		return -ENODATA;
 
-	fts_match_cid(ts_data, ctype[i].type, id_h, id_l, 1);
 	ts_data->ic_info.ids = ctype[i];
 	return 0;
 }
@@ -388,55 +323,6 @@ void fts_release_all_finger(void)
 	mutex_unlock(&ts_data->report_mutex);
 }
 
-/*****************************************************************************
-* Name: fts_input_report_key
-* Brief: process key events,need report key-event if key enable.
-*		if point's coordinate is in (x_dim-50,y_dim-50) ~ (x_dim+50,y_dim+50),
-*		need report it to key event.
-*		x_dim: parse from dts, means key x_coordinate, dimension:+-50
-*		y_dim: parse from dts, means key y_coordinate, dimension:+-50
-* Input:
-* Output:
-* Return: return 0 if it's key event, otherwise return error code
-*****************************************************************************/
-static int fts_input_report_key(struct fts_ts_data *ts_data,
-				struct ts_event *kevent)
-{
-	int i = 0;
-	int x = kevent->x;
-	int y = kevent->y;
-	int *x_dim = &ts_data->pdata->key_x_coords[0];
-	int *y_dim = &ts_data->pdata->key_y_coords[0];
-
-	if (!ts_data->pdata->have_key)
-		return -EINVAL;
-
-	for (i = 0; i < ts_data->pdata->key_number; i++) {
-		if ((x >= x_dim[i] - FTS_KEY_DIM) &&
-		    (x <= x_dim[i] + FTS_KEY_DIM) &&
-		    (y >= y_dim[i] - FTS_KEY_DIM) &&
-		    (y <= y_dim[i] + FTS_KEY_DIM)) {
-			if (EVENT_DOWN(kevent->flag) &&
-			    !(ts_data->key_state & (1 << i))) {
-				input_report_key(ts_data->input_dev,
-						 ts_data->pdata->keys[i], 1);
-				ts_data->key_state |= (1 << i);
-				dev_dbg(ts_data->dev, "Key%d(%d,%d) DOWN!", i,
-					x, y);
-			} else if (EVENT_UP(kevent->flag) &&
-				   (ts_data->key_state & (1 << i))) {
-				input_report_key(ts_data->input_dev,
-						 ts_data->pdata->keys[i], 0);
-				ts_data->key_state &= ~(1 << i);
-				dev_dbg(ts_data->dev, "Key%d(%d,%d) Up!", i, x,
-					y);
-			}
-			return 0;
-		}
-	}
-	return -EINVAL;
-}
-
 static int fts_input_report_b(struct fts_ts_data *ts_data,
 			      struct ts_event *events)
 {
@@ -448,9 +334,6 @@ static int fts_input_report_b(struct fts_ts_data *ts_data,
 	struct input_dev *input_dev = ts_data->input_dev;
 
 	for (i = 0; i < ts_data->touch_event_num; i++) {
-		if (fts_input_report_key(ts_data, &events[i]) == 0)
-			continue;
-
 		touch_event_coordinate = true;
 		if (EVENT_DOWN(events[i].flag)) {
 			input_mt_slot(input_dev, events[i].id);
@@ -468,9 +351,7 @@ static int fts_input_report_b(struct fts_ts_data *ts_data,
 			touch_down_point_cur |= (1 << events[i].id);
 			touch_point_pre |= (1 << events[i].id);
 
-			if ((ts_data->log_level >= 2) ||
-			    ((1 == ts_data->log_level) &&
-			     (FTS_TOUCH_DOWN == events[i].flag))) {
+			if (FTS_TOUCH_DOWN == events[i].flag) {
 				dev_dbg(ts_data->dev,
 					"[B]P%d(%d, %d)[p:%d,tm:%d] DOWN!",
 					events[i].id, events[i].x, events[i].y,
@@ -481,9 +362,7 @@ static int fts_input_report_b(struct fts_ts_data *ts_data,
 			input_mt_report_slot_state(input_dev, MT_TOOL_FINGER,
 						   false);
 			touch_point_pre &= ~(1 << events[i].id);
-			if (ts_data->log_level >= 1)
-				dev_dbg(ts_data->dev, "[B]P%d UP!",
-					events[i].id);
+			dev_dbg(ts_data->dev, "[B]P%d UP!", events[i].id);
 		}
 	}
 
@@ -491,8 +370,7 @@ static int fts_input_report_b(struct fts_ts_data *ts_data,
 		for (i = 0; i < max_touch_num; i++) {
 			if ((1 << i) &
 			    (touch_point_pre ^ touch_down_point_cur)) {
-				if (ts_data->log_level >= 1)
-					dev_dbg(ts_data->dev, "[B]P%d UP!", i);
+				dev_dbg(ts_data->dev, "[B]P%d UP!", i);
 				input_mt_slot(input_dev, i);
 				input_mt_report_slot_state(
 					input_dev, MT_TOOL_FINGER, false);
@@ -503,7 +381,7 @@ static int fts_input_report_b(struct fts_ts_data *ts_data,
 	if (touch_down_point_cur)
 		input_report_key(input_dev, BTN_TOUCH, 1);
 	else if (touch_event_coordinate || ts_data->touch_points) {
-		if (ts_data->touch_points && (ts_data->log_level >= 1))
+		if (ts_data->touch_points)
 			dev_dbg(ts_data->dev, "[B]Points All Up!");
 		input_report_key(input_dev, BTN_TOUCH, 0);
 	}
@@ -550,8 +428,7 @@ static int fts_read_parse_touchdata(struct fts_ts_data *ts_data, u8 *touch_buf)
 		return TOUCH_ERROR;
 	}
 
-	if (ts_data->log_level >= 3)
-		fts_show_touch_buffer(touch_buf, ts_data->ta_size);
+	fts_show_touch_buffer(touch_buf, ts_data->ta_size);
 
 	if (ret)
 		return TOUCH_IGNORE;
@@ -682,57 +559,9 @@ static int fts_irq_read_report(struct fts_ts_data *ts_data)
 		mutex_unlock(&ts_data->report_mutex);
 		break;
 
-	case TOUCH_EXTRA_MSG:
-		if (!ts_data->touch_analysis_support) {
-			dev_err(ts_data->dev, "touch_analysis is disabled");
-			return -EINVAL;
-		}
-
-		event_num = touch_buf[FTS_TOUCH_E_NUM] & 0x0F;
-		if (!event_num || (event_num > max_touch_num)) {
-			dev_err(ts_data->dev, "invalid touch event num(%d)",
-				event_num);
-			return -EIO;
-		}
-
-		ts_data->touch_event_num = event_num;
-		for (i = 0; i < event_num; i++) {
-			base = FTS_ONE_TCH_LEN * i + 4;
-			pointid = (touch_buf[FTS_TOUCH_OFF_ID_YH + base]) >> 4;
-			if (pointid >= max_touch_num) {
-				dev_err(ts_data->dev,
-					"touch point ID(%d) beyond max_touch_number(%d)",
-					pointid, max_touch_num);
-				return -EINVAL;
-			}
-
-			events[i].id = pointid;
-			events[i].flag = touch_buf[FTS_TOUCH_OFF_E_XH + base] >>
-					 6;
-			events[i].x =
-				((touch_buf[FTS_TOUCH_OFF_E_XH + base] & 0x0F)
-				 << 8) +
-				(touch_buf[FTS_TOUCH_OFF_XL + base] & 0xFF);
-			events[i].y =
-				((touch_buf[FTS_TOUCH_OFF_ID_YH + base] & 0x0F)
-				 << 8) +
-				(touch_buf[FTS_TOUCH_OFF_YL + base] & 0xFF);
-			events[i].p = touch_buf[FTS_TOUCH_OFF_PRE + base];
-			events[i].area = touch_buf[FTS_TOUCH_OFF_AREA + base];
-			if (events[i].p <= 0)
-				events[i].p = 0x3F;
-			if (events[i].area <= 0)
-				events[i].area = 0x09;
-		}
-
-		mutex_lock(&ts_data->report_mutex);
-		fts_input_report_b(ts_data, events);
-		mutex_unlock(&ts_data->report_mutex);
-		break;
-
 	case TOUCH_FW_INIT:
 		fts_release_all_finger();
-		fts_tp_state_recovery(ts_data);
+		fts_wait_tp_to_valid();
 		break;
 
 	case TOUCH_IGNORE:
@@ -750,29 +579,9 @@ static int fts_irq_read_report(struct fts_ts_data *ts_data)
 static irqreturn_t fts_irq_handler(int irq, void *data)
 {
 	struct fts_ts_data *ts_data = fts_data;
-#if defined(CONFIG_PM) && FTS_PATCH_COMERR_PM
-	int ret = 0;
 
-	if ((ts_data->suspended) && (ts_data->pm_suspend)) {
-		ret = wait_for_completion_timeout(
-			&ts_data->pm_completion,
-			msecs_to_jiffies(FTS_TIMEOUT_COMERR_PM));
-		if (!ret) {
-			dev_err(ts_data->dev,
-				"Bus don't resume from pm(deep),timeout,skip irq");
-			return IRQ_HANDLED;
-		}
-	}
-#endif
 	ts_data->intr_jiffies = jiffies;
 	fts_irq_read_report(ts_data);
-	if (ts_data->touch_analysis_support && ts_data->ta_flag) {
-		ts_data->ta_flag = 0;
-		if (ts_data->ta_buf && ts_data->ta_size)
-			memcpy(ts_data->ta_buf, ts_data->touch_buf,
-			       ts_data->ta_size);
-		wake_up_interruptible(&ts_data->ts_waitqueue);
-	}
 
 	return IRQ_HANDLED;
 }
@@ -796,7 +605,6 @@ static int fts_irq_registration(struct fts_ts_data *ts_data)
 static int fts_input_init(struct fts_ts_data *ts_data)
 {
 	int ret = 0;
-	int key_num = 0;
 	struct fts_ts_platform_data *pdata = ts_data->pdata;
 	struct input_dev *input_dev;
 
@@ -822,13 +630,6 @@ static int fts_input_init(struct fts_ts_data *ts_data)
 	__set_bit(EV_KEY, input_dev->evbit);
 	__set_bit(BTN_TOUCH, input_dev->keybit);
 	__set_bit(INPUT_PROP_DIRECT, input_dev->propbit);
-
-	if (pdata->have_key) {
-		dev_info(ts_data->dev, "set key capabilities");
-		for (key_num = 0; key_num < pdata->key_number; key_num++)
-			input_set_capability(input_dev, EV_KEY,
-					     pdata->keys[key_num]);
-	}
 
 	input_mt_init_slots(input_dev, pdata->max_touch_number,
 			    INPUT_MT_DIRECT);
@@ -861,10 +662,6 @@ static int fts_buffer_init(struct fts_ts_data *ts_data)
 		return -ENOMEM;
 
 	ts_data->touch_size = FTS_TOUCH_DATA_LEN;
-
-	ts_data->touch_analysis_support = 0;
-	ts_data->ta_flag = 0;
-	ts_data->ta_size = 0;
 
 	return 0;
 }
@@ -984,43 +781,6 @@ static int fts_parse_dt(struct device *dev, struct fts_ts_platform_data *pdata)
 	if (ret < 0)
 		dev_err(dev, "Unable to get display-coords");
 
-	/* key */
-	pdata->have_key = of_property_read_bool(np, "focaltech,have-key");
-	if (pdata->have_key) {
-		ret = of_property_read_u32(np, "focaltech,key-number",
-					   &pdata->key_number);
-		if (ret < 0)
-			dev_err(dev, "Key number undefined!");
-
-		ret = of_property_read_u32_array(
-			np, "focaltech,keys", pdata->keys, pdata->key_number);
-		if (ret < 0)
-			dev_err(dev, "Keys undefined!");
-		else if (pdata->key_number > FTS_MAX_KEYS)
-			pdata->key_number = FTS_MAX_KEYS;
-
-		ret = of_property_read_u32_array(np, "focaltech,key-x-coords",
-						 pdata->key_x_coords,
-						 pdata->key_number);
-		if (ret < 0)
-			dev_err(dev, "Key Y Coords undefined!");
-
-		ret = of_property_read_u32_array(np, "focaltech,key-y-coords",
-						 pdata->key_y_coords,
-						 pdata->key_number);
-		if (ret < 0)
-			dev_err(dev, "Key X Coords undefined!");
-
-		dev_info(dev,
-			 "VK Number:%d, key:(%d,%d,%d), "
-			 "coords:(%d,%d),(%d,%d),(%d,%d)",
-			 pdata->key_number, pdata->keys[0], pdata->keys[1],
-			 pdata->keys[2], pdata->key_x_coords[0],
-			 pdata->key_y_coords[0], pdata->key_x_coords[1],
-			 pdata->key_y_coords[1], pdata->key_x_coords[2],
-			 pdata->key_y_coords[2]);
-	}
-
 	pdata->irq_gpio = of_get_named_gpio(np, "focaltech,irq-gpio", 0);
 	if (pdata->irq_gpio < 0)
 		dev_err(dev, "Unable to get irq_gpio");
@@ -1032,8 +792,8 @@ static int fts_parse_dt(struct device *dev, struct fts_ts_platform_data *pdata)
 		pdata->max_touch_number = FTS_MAX_POINTS_SUPPORT;
 	} else {
 		if (temp_val < 2)
-			pdata->max_touch_number =
-				2; /* max_touch_number must >= 2 */
+			/* max_touch_number must >= 2 */
+			pdata->max_touch_number = 2;
 		else if (temp_val > FTS_MAX_POINTS_SUPPORT)
 			pdata->max_touch_number = FTS_MAX_POINTS_SUPPORT;
 		else
@@ -1091,8 +851,7 @@ static int fts_ts_resume(struct device *dev)
 	struct fts_ts_data *ts_data = fts_data;
 	struct input_dev *input_dev = ts_data->input_dev;
 
-	dev_info(dev, "down flag = %d,blank flag = %d",
-		 ts_data->fod_info.fp_down_report, ts_data->blank_up);
+	dev_info(dev, "blank flag = %d", ts_data->blank_up);
 	if ((ts_data->fod_info.fp_down_report) && (ts_data->blank_up)) {
 		ts_data->fod_info.fp_down_report = 0;
 		input_sync(input_dev);
@@ -1110,38 +869,10 @@ static int fts_ts_resume(struct device *dev)
 		fts_power_resume(ts_data);
 
 	fts_wait_tp_to_valid();
-	fts_ex_mode_recovery(ts_data);
 
 	ts_data->blank_up = 0;
 	return 0;
 }
-
-#if defined(CONFIG_PM) && FTS_PATCH_COMERR_PM
-static int fts_pm_suspend(struct device *dev)
-{
-	struct fts_ts_data *ts_data = dev_get_drvdata(dev);
-
-	dev_info(dev, "system enters into pm_suspend");
-	ts_data->pm_suspend = true;
-	reinit_completion(&ts_data->pm_completion);
-	return 0;
-}
-
-static int fts_pm_resume(struct device *dev)
-{
-	struct fts_ts_data *ts_data = dev_get_drvdata(dev);
-
-	dev_info(dev, "system resumes from pm_suspend");
-	ts_data->pm_suspend = false;
-	complete(&ts_data->pm_completion);
-	return 0;
-}
-
-static const struct dev_pm_ops fts_dev_pm_ops = {
-	.suspend = fts_pm_suspend,
-	.resume = fts_pm_resume,
-};
-#endif
 
 static int fts_ts_probe(struct spi_device *spi)
 {
@@ -1184,10 +915,15 @@ static int fts_ts_probe(struct spi_device *spi)
 	if (ret)
 		return dev_err_probe(&spi->dev, ret, "Failed power on\n");
 
+	/* Get firmware path */
+	ret = of_property_read_string(spi->dev.of_node, "firmware-name", &ts_data->firmware_path);
+	if (ret)
+		return dev_err_probe(&spi->dev, ret,
+				     "Failed to read firmware-name property\n");
+
 	fts_data = ts_data;
 	ts_data->spi = spi;
 	ts_data->dev = &spi->dev;
-	ts_data->log_level = 1;
 
 	ts_data->bus_type = BUS_TYPE_SPI_V2;
 	spi_set_drvdata(spi, ts_data);
@@ -1247,14 +983,6 @@ static int fts_ts_probe(struct spi_device *spi)
 		goto err_irq_req;
 	}
 
-	ret = fts_create_sysfs(ts_data);
-	if (ret)
-		dev_err(&spi->dev, "create sysfs node fail");
-
-	ret = fts_ex_mode_init(ts_data);
-	if (ret)
-		dev_err(&spi->dev, "init glove/cover/charger fail");
-
 	ret = fts_irq_registration(ts_data);
 	if (ret) {
 		dev_err(&spi->dev, "request irq failed");
@@ -1267,11 +995,6 @@ static int fts_ts_probe(struct spi_device *spi)
 
 	if (ts_data->ts_workqueue)
 		INIT_WORK(&ts_data->resume_work, fts_resume_work);
-
-#if defined(CONFIG_PM) && FTS_PATCH_COMERR_PM
-	init_completion(&ts_data->pm_completion);
-	ts_data->pm_suspend = false;
-#endif
 
 	device_init_wakeup(ts_data->dev, true);
 
@@ -1297,15 +1020,9 @@ err_bus_init:
 static void fts_ts_remove(struct spi_device *spi)
 {
 	cancel_work_sync(&fts_data->resume_work);
-	fts_remove_sysfs(fts_data);
-	fts_ex_mode_exit(fts_data);
-
 	fts_fwupg_exit(fts_data);
-
 	free_irq(fts_data->irq, fts_data);
-
 	fts_bus_exit(fts_data);
-
 	input_unregister_device(fts_data->input_dev);
 
 	if (fts_data->ts_workqueue)
@@ -1318,14 +1035,14 @@ static void fts_ts_remove(struct spi_device *spi)
 }
 
 static const struct spi_device_id fts_ts_id[] = {
-	{ FTS_DRIVER_NAME, 0 },
-	{},
+	{ "ft3680" },
+	{ },
 };
 MODULE_DEVICE_TABLE(spi, fts_ts_id);
 
 static const struct of_device_id fts_dt_match[] = {
-	{ .compatible = "focaltech,fts_ts", },
-	{},
+	{ .compatible = "focaltech,ft3680", },
+	{ },
 };
 MODULE_DEVICE_TABLE(of, fts_dt_match);
 
@@ -1335,9 +1052,6 @@ static struct spi_driver fts_ts_driver = {
 	.driver = {
 		.name = FTS_DRIVER_NAME,
 		.owner = THIS_MODULE,
-#if defined(CONFIG_PM) && FTS_PATCH_COMERR_PM
-		.pm = &fts_dev_pm_ops,
-#endif
 		.of_match_table = of_match_ptr(fts_dt_match),
 	},
 	.id_table = fts_ts_id,
