--- drivers/input/touchscreen/focaltech_touch/focaltech_ex_mode.c
@@ -1,338 +1,0 @@
-/*
- * FocalTech ftxxxx TouchScreen driver.
- *
- * Copyright (c) 2012-2020, Focaltech Ltd. All rights reserved.
- *
- * This software is licensed under the terms of the GNU General Public
- * License version 2, as published by the Free Software Foundation, and
- * may be copied, distributed, and modified under those terms.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- */
-
-#include "focaltech_core.h"
-
-enum _ex_mode {
-	MODE_GLOVE = 0,
-	MODE_COVER,
-	MODE_CHARGER,
-	MODE_EDGE,
-};
-
-static int fts_ex_mode_switch(enum _ex_mode mode, u8 value)
-{
-	int ret = 0;
-	u8 m_val = 0;
-
-	if (value)
-		m_val = value;
-	else
-		m_val = 0x00;
-
-	switch (mode) {
-	case MODE_GLOVE:
-		ret = fts_write_reg(FTS_REG_GLOVE_MODE_EN, m_val);
-		if (ret < 0)
-			dev_err(fts_data->dev, "MODE_GLOVE switch to %d fail",
-				m_val);
-		break;
-
-	case MODE_EDGE:
-		ret = fts_write_reg(FTS_REG_EDGE_MODE_EN, m_val);
-		if (ret < 0)
-			dev_err(fts_data->dev, "MODE_EDGE swith to %d fail",
-				m_val);
-		break;
-	case MODE_COVER:
-		ret = fts_write_reg(FTS_REG_COVER_MODE_EN, m_val);
-		if (ret < 0)
-			dev_err(fts_data->dev, "MODE_COVER switch to %d fail",
-				m_val);
-		break;
-	case MODE_CHARGER:
-		ret = fts_write_reg(FTS_REG_CHARGER_MODE_EN, m_val);
-		if (ret < 0)
-			dev_err(fts_data->dev, "MODE_CHARGER switch to %d fail",
-				m_val);
-		break;
-	default:
-		dev_err(fts_data->dev, "mode(%d) unsupport", mode);
-		ret = -EINVAL;
-		break;
-	}
-
-	return ret;
-}
-
-static ssize_t fts_edge_mode_show(struct device *dev,
-				  struct device_attribute *attr, char *buf)
-{
-	int count = 0;
-	u8 val = 0;
-	struct fts_ts_data *ts_data = fts_data;
-	struct input_dev *input_dev = ts_data->input_dev;
-
-	mutex_lock(&input_dev->mutex);
-	fts_read_reg(FTS_REG_EDGE_MODE_EN, &val);
-	count = snprintf(buf + count, PAGE_SIZE, "Edge Mode:%s\n",
-			 ts_data->edge_mode ? "On" : "Off");
-	count += snprintf(buf + count, PAGE_SIZE, "Edge Reg(0x8C):%d\n", val);
-	mutex_unlock(&input_dev->mutex);
-
-	return count;
-}
-
-static ssize_t fts_edge_mode_store(struct device *dev,
-				   struct device_attribute *attr,
-				   const char *buf, size_t count)
-{
-	int ret = 0;
-	struct fts_ts_data *ts_data = fts_data;
-
-	if (FTS_SYSFS_ECHO_ON(buf)) {
-		if (!ts_data->edge_mode) {
-			dev_dbg(fts_data->dev, "enter edge mode");
-			if (buf[0] == '1') /*USB in right*/ {
-				ret = fts_ex_mode_switch(MODE_EDGE, 1);
-				if (ret >= 0) {
-					ts_data->edge_mode = 1;
-				}
-			} else if (buf[0] == '2') /*USB in left*/ {
-				ret = fts_ex_mode_switch(MODE_EDGE, 2);
-				if (ret >= 0) {
-					ts_data->edge_mode = 2;
-				}
-			}
-		}
-	} else if (FTS_SYSFS_ECHO_OFF(buf)) {
-		if (ts_data->edge_mode) {
-			dev_dbg(fts_data->dev, "exit edge mode");
-			ret = fts_ex_mode_switch(MODE_EDGE, 0);
-			if (ret >= 0) {
-				ts_data->edge_mode = DISABLE;
-			}
-		}
-	}
-
-	dev_dbg(fts_data->dev, "edge mode:%d", ts_data->edge_mode);
-	return count;
-}
-
-static ssize_t fts_glove_mode_show(struct device *dev,
-				   struct device_attribute *attr, char *buf)
-{
-	int count = 0;
-	u8 val = 0;
-	struct fts_ts_data *ts_data = dev_get_drvdata(dev);
-	struct input_dev *input_dev = ts_data->input_dev;
-
-	mutex_lock(&input_dev->mutex);
-	fts_read_reg(FTS_REG_GLOVE_MODE_EN, &val);
-	count = snprintf(buf + count, PAGE_SIZE, "Glove Mode:%s\n",
-			 ts_data->glove_mode ? "On" : "Off");
-	count += snprintf(buf + count, PAGE_SIZE, "Glove Reg(0xC0):%d\n", val);
-	mutex_unlock(&input_dev->mutex);
-
-	return count;
-}
-
-static ssize_t fts_glove_mode_store(struct device *dev,
-				    struct device_attribute *attr,
-				    const char *buf, size_t count)
-{
-	int ret = 0;
-	struct fts_ts_data *ts_data = dev_get_drvdata(dev);
-
-	if (FTS_SYSFS_ECHO_ON(buf)) {
-		if (!ts_data->glove_mode) {
-			dev_dbg(fts_data->dev, "enter glove mode");
-			ret = fts_ex_mode_switch(MODE_GLOVE, ENABLE);
-			if (ret >= 0) {
-				ts_data->glove_mode = ENABLE;
-			}
-		}
-	} else if (FTS_SYSFS_ECHO_OFF(buf)) {
-		if (ts_data->glove_mode) {
-			dev_dbg(fts_data->dev, "exit glove mode");
-			ret = fts_ex_mode_switch(MODE_GLOVE, DISABLE);
-			if (ret >= 0) {
-				ts_data->glove_mode = DISABLE;
-			}
-		}
-	}
-
-	dev_dbg(fts_data->dev, "glove mode:%d", ts_data->glove_mode);
-	return count;
-}
-
-static ssize_t fts_cover_mode_show(struct device *dev,
-				   struct device_attribute *attr, char *buf)
-{
-	int count = 0;
-	u8 val = 0;
-	struct fts_ts_data *ts_data = dev_get_drvdata(dev);
-	struct input_dev *input_dev = ts_data->input_dev;
-
-	mutex_lock(&input_dev->mutex);
-	fts_read_reg(FTS_REG_COVER_MODE_EN, &val);
-	count = snprintf(buf + count, PAGE_SIZE, "Cover Mode:%s\n",
-			 ts_data->cover_mode ? "On" : "Off");
-	count += snprintf(buf + count, PAGE_SIZE, "Cover Reg(0xC1):%d\n", val);
-	mutex_unlock(&input_dev->mutex);
-
-	return count;
-}
-
-static ssize_t fts_cover_mode_store(struct device *dev,
-				    struct device_attribute *attr,
-				    const char *buf, size_t count)
-{
-	int ret = 0;
-	struct fts_ts_data *ts_data = dev_get_drvdata(dev);
-
-	if (FTS_SYSFS_ECHO_ON(buf)) {
-		if (!ts_data->cover_mode) {
-			dev_dbg(fts_data->dev, "enter cover mode");
-			ret = fts_ex_mode_switch(MODE_COVER, ENABLE);
-			if (ret >= 0) {
-				ts_data->cover_mode = ENABLE;
-			}
-		}
-	} else if (FTS_SYSFS_ECHO_OFF(buf)) {
-		if (ts_data->cover_mode) {
-			dev_dbg(fts_data->dev, "exit cover mode");
-			ret = fts_ex_mode_switch(MODE_COVER, DISABLE);
-			if (ret >= 0) {
-				ts_data->cover_mode = DISABLE;
-			}
-		}
-	}
-
-	dev_dbg(fts_data->dev, "cover mode:%d", ts_data->cover_mode);
-	return count;
-}
-
-static ssize_t fts_charger_mode_show(struct device *dev,
-				     struct device_attribute *attr, char *buf)
-{
-	int count = 0;
-	u8 val = 0;
-	struct fts_ts_data *ts_data = dev_get_drvdata(dev);
-	struct input_dev *input_dev = ts_data->input_dev;
-
-	mutex_lock(&input_dev->mutex);
-	fts_read_reg(FTS_REG_CHARGER_MODE_EN, &val);
-	count = snprintf(buf + count, PAGE_SIZE, "Charger Mode:%s\n",
-			 ts_data->charger_mode ? "On" : "Off");
-	count +=
-		snprintf(buf + count, PAGE_SIZE, "Charger Reg(0x8B):%d\n", val);
-	mutex_unlock(&input_dev->mutex);
-
-	return count;
-}
-
-static ssize_t fts_charger_mode_store(struct device *dev,
-				      struct device_attribute *attr,
-				      const char *buf, size_t count)
-{
-	int ret = 0;
-	struct fts_ts_data *ts_data = dev_get_drvdata(dev);
-
-	if (FTS_SYSFS_ECHO_ON(buf)) {
-		if (!ts_data->charger_mode) {
-			dev_dbg(fts_data->dev, "enter charger mode");
-			ret = fts_ex_mode_switch(MODE_CHARGER, ENABLE);
-			if (ret >= 0) {
-				ts_data->charger_mode = ENABLE;
-			}
-		}
-	} else if (FTS_SYSFS_ECHO_OFF(buf)) {
-		if (ts_data->charger_mode) {
-			dev_dbg(fts_data->dev, "exit charger mode");
-			ret = fts_ex_mode_switch(MODE_CHARGER, DISABLE);
-			if (ret >= 0) {
-				ts_data->charger_mode = DISABLE;
-			}
-		}
-	}
-
-	dev_dbg(fts_data->dev, "charger mode:%d", ts_data->glove_mode);
-	return count;
-}
-
-/* read and write charger mode
- * read example: cat fts_glove_mode        ---read  glove mode
- * write example:echo 1 > fts_glove_mode   ---write glove mode to 01
- */
-static DEVICE_ATTR(fts_glove_mode, S_IRUGO | S_IWUSR, fts_glove_mode_show,
-		   fts_glove_mode_store);
-
-static DEVICE_ATTR(fts_cover_mode, S_IRUGO | S_IWUSR, fts_cover_mode_show,
-		   fts_cover_mode_store);
-
-static DEVICE_ATTR(fts_charger_mode, S_IRUGO | S_IWUSR, fts_charger_mode_show,
-		   fts_charger_mode_store);
-static DEVICE_ATTR(fts_edge_mode, S_IRUGO | S_IWUSR, fts_edge_mode_show,
-		   fts_edge_mode_store);
-
-static struct attribute *fts_touch_mode_attrs[] = {
-	&dev_attr_fts_glove_mode.attr,
-	&dev_attr_fts_cover_mode.attr,
-	&dev_attr_fts_charger_mode.attr,
-	&dev_attr_fts_edge_mode.attr,
-	NULL,
-};
-
-static struct attribute_group fts_touch_mode_group = {
-	.attrs = fts_touch_mode_attrs,
-};
-
-int fts_ex_mode_recovery(struct fts_ts_data *ts_data)
-{
-	if (ts_data->glove_mode) {
-		fts_ex_mode_switch(MODE_GLOVE, ENABLE);
-	}
-
-	if (ts_data->cover_mode) {
-		fts_ex_mode_switch(MODE_COVER, ENABLE);
-	}
-
-	if (ts_data->charger_mode) {
-		fts_ex_mode_switch(MODE_CHARGER, ENABLE);
-	}
-
-	if (ts_data->edge_mode) {
-		fts_ex_mode_switch(MODE_EDGE, ENABLE);
-	}
-	return 0;
-}
-
-int fts_ex_mode_init(struct fts_ts_data *ts_data)
-{
-	int ret = 0;
-
-	ts_data->glove_mode = DISABLE;
-	ts_data->cover_mode = DISABLE;
-	ts_data->charger_mode = DISABLE;
-	ts_data->edge_mode = DISABLE;
-
-	ret = sysfs_create_group(&ts_data->dev->kobj, &fts_touch_mode_group);
-	if (ret < 0) {
-		dev_err(fts_data->dev, "create sysfs(ex_mode) fail");
-		sysfs_remove_group(&ts_data->dev->kobj, &fts_touch_mode_group);
-		return ret;
-	} else {
-		dev_dbg(fts_data->dev, "create sysfs(ex_mode) successfully");
-	}
-
-	return 0;
-}
-
-int fts_ex_mode_exit(struct fts_ts_data *ts_data)
-{
-	sysfs_remove_group(&ts_data->dev->kobj, &fts_touch_mode_group);
-	return 0;
-}
